+++ documentation/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## yyyy.mm.dd - version 1.2.0
 
+* Add the Environment class to identify the environment where the addon is 
+running
 * Allow classes to define constants with the Arr:freeze() method
 
 ## 2024.04.25 - version 1.1.0